Replacing files with correct version
There was some confusion about which was the most up-to-date version of the neuroblinks code. The version previously in the master branch works and can collect data, but it does not collect data from the rotary encoders. So, I have copied files from the neuroblinks folder that Alvaro has actually been using that should collect data from the rotary encoders.