Modified GUI, updated corresponding code
Made GUI slightly more intuitively usable (got rid of boxes that don't do anything, indicated what things you usually don't want to touch during the session) and reorganized it a bit to get rid of the empty space. Also commented out the line in Launch that was forcing the GUI window to appear with dimensions for the old version.